Move resolve test to swarm_dial_test
